requirements.txt
#Google pubsub
google-cloud-pubsub==0.39.1
